refactor - seperate code into components
