+++ src/components/AddTask.module.scss
@@ -1,0 +1,28 @@
+.newTask {
+  display: flex;
+  gap: 10px;
+
+  .taskName {
+    border-bottom: 1px solid var(--color-light-gray);
+    background: inherit;
+    width: 140px;
+    padding-left: 5px;
+
+    &::placeholder {
+      font-size: 10px;
+    }
+  }
+
+  .taskName:invalid {
+    border-bottom: 1px solid var(--color-danger);
+  }
+
+  .newTaskAddButton {
+    border: none;
+    background: var(--color-blue);
+    border-radius: 5px;
+    min-width: 85px;
+    height: 30px;
+    color: white;
+  }
+}
