+++ src/components/MainPage.module.scss
@@ -13,136 +13,3 @@
   flex-direction: column;
   gap: 10px;
 }
-
-.newTask {
-  display: flex;
-  gap: 10px;
-
-  .taskName {
-    border-bottom: 1px solid var(--color-light-gray);
-    background: inherit;
-    // font-size: 10px;
-    width: 140px;
-    padding-left: 5px;
-
-    &::placeholder {
-      font-size: 10px;
-    }
-  }
-
-  .taskName:invalid {
-    border-bottom: 1px solid var(--color-danger);
-  }
-
-  .newTaskAddButton {
-    border: none;
-    background: var(--color-blue);
-    border-radius: 5px;
-    min-width: 85px;
-    height: 30px;
-    color: white;
-  }
-}
-
-.filters {
-  display: flex;
-  gap: 10px;
-  
-  * {
-    background: none;
-    color: var(--color-dark-gray);
-    font-size: 15px;
-  }
-
-  .active {
-    color: var(--color-blue);
-    font-weight: bold;
-    border-bottom: 2px solid var(--color-blue);
-  }
-}
-
-.taskWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  width: 225px;
-  height: 35px;
-  border-radius: 5px;
-  background: white;
-
-  .checkboxWrapper {
-    position: relative;
-
-    .checkbox {
-      appearance: none;
-      width: 13px;
-      height: 13px;
-      border-radius: 50%;
-      border: 1px solid var(--color-light-gray);
-      position: absolute;
-      transform: translate(-50%, -50%);
-      top: 0;
-      left: 0;
-      
-      &:checked {
-        background-color: var(--color-blue);
-        background-image: url('../assets/check-icon.svg');
-        background-size: 70%;
-        background-position: 50%;
-        background-repeat: no-repeat;
-      }
-  
-      &:hover {
-        cursor: pointer;
-      }
-    }
-  
-    .checkIcon {
-      width: 7px;
-      height: 7px;
-      position: absolute;
-      transform: translate(-50%, -50%);
-      top: 0;
-      left: 0;
-    }
-  }
-
-  .taskName,
-  .taskNameEdit,
-  .taskIsDone {
-    width: 100px;
-    overflow: hidden;
-    margin-left: 4px;
-  }
-
-  .taskNameEdit:invalid {
-    border: 1px solid red;
-    border-radius: 2px;
-  }
-
-  .taskIsDone {
-    color: var(--color-dark-gray);
-    text-decoration: line-through;
-  }
-
-  .taskEditButton,
-  .taskDeleteButton {
-    height: 25px;
-    width: 30px;
-    border-radius: 3px;
-    background: var(--color-blue);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    * {
-      width: 10px;
-      height: 10px;
-    }
-  }
-
-  .taskDeleteButton {
-    background: var(--color-danger);
-  }
-}
