+++ src/components/TodoItem.module.scss
@@ -1,0 +1,85 @@
+.taskWrapper {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  width: 225px;
+  height: 35px;
+  border-radius: 5px;
+  background: white;
+
+  .checkboxWrapper {
+    position: relative;
+
+    .checkbox {
+      appearance: none;
+      width: 13px;
+      height: 13px;
+      border-radius: 50%;
+      border: 1px solid var(--color-light-gray);
+      position: absolute;
+      transform: translate(-50%, -50%);
+      top: 0;
+      left: 0;
+      
+      &:checked {
+        background-color: var(--color-blue);
+        background-image: url('../assets/check-icon.svg');
+        background-size: 70%;
+        background-position: 50%;
+        background-repeat: no-repeat;
+      }
+  
+      &:hover {
+        cursor: pointer;
+      }
+    }
+  
+    .checkIcon {
+      width: 7px;
+      height: 7px;
+      position: absolute;
+      transform: translate(-50%, -50%);
+      top: 0;
+      left: 0;
+    }
+  }
+
+  .taskName,
+  .taskNameEdit,
+  .taskIsDone {
+    width: 100px;
+    overflow: hidden;
+    margin-left: 4px;
+  }
+
+  .taskNameEdit:invalid {
+    border: 1px solid red;
+    border-radius: 2px;
+  }
+
+  .taskIsDone {
+    color: var(--color-dark-gray);
+    text-decoration: line-through;
+  }
+
+  .taskEditButton,
+  .taskDeleteButton {
+    height: 25px;
+    width: 30px;
+    border-radius: 3px;
+    background: var(--color-blue);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    * {
+      width: 10px;
+      height: 10px;
+    }
+  }
+
+  .taskDeleteButton {
+    background: var(--color-danger);
+  }
+}
